Python/doctr_requirements.txt
python-doctr
torch
torchvision